+++ CHANGELOG.md
@@ -6,6 +6,13 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+## [0.2.0/0.2.2] - 2018-10-22
+### Added
+- AMRedux created to deal with default redux actions and dispatches
+- Default populate defined on config
+### Changed
+- Setting empty object on statics of AMRedux
+
 ## [0.1.3] - 2018-10-18
 ### Added
 - Relationship methods on synchronization